pyproject.toml
[tool.poetry]
name = "user-management-micro-service"
version = "0.1.0"
description = ""
authors = ["Artyom Vorozhbyanov <artyom.vorozhbyanov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.6"
sqlalchemy = "2.0.23"
alembic = "1.12.1"
fastapi = "0.104.1"
uvicorn = "0.24.0.post1"
asyncpg = "0.29.0"
pydantic-settings = "2.0.3"
pyjwt = "2.8.0"
python-dateutil = "2.8.2"
pytz = "2023.3.post1"
passlib = "1.7.4"
python-multipart = "0.0.6"
cryptography = "41.0.5"
redis = "5.0.1"


[tool.poetry.group.dev.dependencies]
black = "23.10.1"
isort = "5.12.0"
flake8 = "6.1.0"
flake8-builtins = "2.1.0"
flake8-bandit = "4.1.1"
flake8-comprehensions = "3.14.0"
flake8-fixme = "1.1.1"
flake8-multiline-containers = "0.0.19"
flake8-mutable = "1.2.0"
flake8-print = "5.0.0"
flake8-quotes = "3.3.2"
flake8-simplify = "0.21.0"
pytest-asyncio = "0.21.1"
httpx = "0.25.0"
fakeredis = "2.20.0"

[tool.black]
line-length = 120
target-version = ["py310"]
include = '^.+\.pyi?$'
extend-exclude = '''
(
    .*\.idea.*
    | .*\.vscode.*
    | .*\.git.*
    | .*\.local.*
)
'''

[tool.isort]
line_length = 120
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
